Support custom scheme for QR codes
If people have many variants of the same app (say a development build,
an rc build, and production) then the default scheme that Expo adds to
each app for handling QR previews will be the same:

exp+<slug>://

This means whichever app was installed last is the one that will be
loaded by QR previews, which often isn't what you want. This commit
adds support for providing a custom `scheme` to the continuous delivery
action that modifies the slug used in the QR code so users can have them
open in the correct app.